Separate Promise implementation from polyfilling
